+++ src/App.jsx
@@ -1,15 +1,9 @@
-import GuestFavorite from './components/GuestFavorite/GuestFavorite.jsx'
-import Review from './components/Review/Review.jsx'
-import ReviewMeta from './components/ReviewMeta/ReviewMeta.jsx'
+import Comment from './pages/Comment/Comment.jsx'
 
 function App() {
   return (
     <>
-      <div>
-        <GuestFavorite />
-        <ReviewMeta />
-        <Review />
-      </div>
+      <Comment />
     </>
   )
 }
