+++ src/components/Review/ReviewMoreButton.jsx
@@ -2,7 +2,7 @@ const data = 16
 
 const ReviewMoreButton = () => {
   return (
-    <button className='border border-solid border-black rounded-lg font-semibold text-center px-4 py-2'>
+    <button className='border border-solid border-black rounded-lg font-semibold text-center mt-10 px-4 py-2'>
       후기 {data}개 모두 보기
     </button>
   )
